requirements.txt
fastapi
uvicorn
numpy
Pillow
onnxruntime
torch
torchvision
python-multipart